Update the README, optimize the test package structure, and add a test case.
