--- plugin/hotswap-agent-mybatis-plus-plugin/src/test/java/org/hotswap/agent/plugin/mybatis/plus/PlusUser2.java
@@ -1,65 +1,0 @@
-/*
- * Copyright 2013-2024 the HotswapAgent authors.
- *
- * This file is part of HotswapAgent.
- *
- * HotswapAgent is free software: you can redistribute it and/or modify it
- * under the terms of the GNU General Public License as published by the
- * Free Software Foundation, either version 2 of the License, or (at your
- * option) any later version.
- *
- * HotswapAgent is distributed in the hope that it will be useful, but
- * WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
- * Public License for more details.
- *
- * You should have received a copy of the GNU General Public License along
- * with HotswapAgent. If not, see http://www.gnu.org/licenses/.
- */
-package org.hotswap.agent.plugin.mybatis.plus;
-
-import com.baomidou.mybatisplus.annotation.TableId;
-import com.baomidou.mybatisplus.annotation.TableName;
-
-@TableName("users")
-public class PlusUser2 {
-
-  @TableId
-  private Integer id;
-  private String name1;
-  private String name2;
-
-  private String gender;
-
-  public String getGender() {
-    return gender;
-  }
-
-  public void setGender(String gender) {
-    this.gender = gender;
-  }
-
-  public Integer getId() {
-    return id;
-  }
-
-  public void setId(Integer id) {
-    this.id = id;
-  }
-
-  public String getName1() {
-    return name1;
-  }
-
-  public void setName1(String name1) {
-    this.name1 = name1;
-  }
-
-  public String getName2() {
-    return name2;
-  }
-
-  public void setName2(String name2) {
-    this.name2 = name2;
-  }
-}
