+++ plugin/hotswap-agent-mybatis-plus-plugin/src/test/java/org/hotswap/agent/plugin/mybatisplus/Application.java
@@ -1,4 +1,4 @@
-package org.hotswap.agent.plugin.mybatis.plus;
+package org.hotswap.agent.plugin.mybatisplus;
 
 import org.hotswap.agent.logging.AgentLogger;
 import org.mybatis.spring.annotation.MapperScan;
@@ -13,14 +13,14 @@ import org.springframework.jdbc.datasource.init.ResourceDatabasePopulator;
 import javax.sql.DataSource;
 
 @SpringBootApplication
-@MapperScan("org.hotswap.agent.plugin.mybatis.plus")
-public class PlusApplication {
-    private static AgentLogger LOGGER = AgentLogger.getLogger(PlusApplication.class);
+@MapperScan("org.hotswap.agent.plugin.mybatisplus.mapper")
+public class Application {
+    private static AgentLogger LOGGER = AgentLogger.getLogger(Application.class);
 
     private static ApplicationContext applicationContext;
     public static void main(String[] args) {
         LOGGER.info("Starting MyBatis Plus Application");
-        applicationContext = SpringApplication.run(PlusApplication.class, args);
+        applicationContext = SpringApplication.run(Application.class, args);
     }
 
     @Bean
@@ -29,7 +29,7 @@ public class PlusApplication {
         DataSourceInitializer initializer = new DataSourceInitializer();
         initializer.setDataSource(dataSource);
         ResourceDatabasePopulator populator = new ResourceDatabasePopulator();
-        populator.addScript(new ClassPathResource("org/hotswap/agent/plugin/mybatis/CreateDB.sql"));
+        populator.addScript(new ClassPathResource("org/hotswap/agent/plugin/mybatisplus/CreateDB.sql"));
         initializer.setDatabasePopulator(populator);
         return initializer;
     }
