+++ plugin/hotswap-agent-mybatis-plus-plugin/src/test/java/org/hotswap/agent/plugin/mybatisplus/mapper/PlusMapper.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2013-2024 the HotswapAgent authors.
+ *
+ * This file is part of HotswapAgent.
+ *
+ * HotswapAgent is free software: you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License as published by the
+ * Free Software Foundation, either version 2 of the License, or (at your
+ * option) any later version.
+ *
+ * HotswapAgent is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
+ * Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along
+ * with HotswapAgent. If not, see http://www.gnu.org/licenses/.
+ */
+package org.hotswap.agent.plugin.mybatisplus.mapper;
+
+import com.baomidou.mybatisplus.core.mapper.BaseMapper;
+import org.apache.ibatis.annotations.Param;
+import org.apache.ibatis.annotations.Select;
+import org.hotswap.agent.plugin.mybatisplus.entity.PlusUser;
+
+@org.apache.ibatis.annotations.Mapper
+public interface PlusMapper extends BaseMapper<PlusUser> {
+
+  @Select("select * from users where name1 = #{name1}")
+  PlusUser getUser(@Param("name1") String name1);
+
+  PlusUser getUserXML(String name);
+
+}
