+++ plugin/hotswap-agent-mybatis-plus-plugin/src/test/resources/application-plus.properties
@@ -9,8 +9,8 @@ spring.h2.console.enabled=true
 spring.h2.console.path=/h2
 spring.h2.console.settings.web-allow-others=true
 
-mybatis-plus.mapper-locations=classpath:org/hotswap/agent/plugin/mybatis/*Mapper.xml
+mybatis-plus.mapper-locations=classpath:org/hotswap/agent/plugin/mybatisplus/*Mapper.xml
 
 # log sql
 logging.level.org.apache.ibatis=debug
-logging.level.org.hotswap.agent.plugin.mybatis.plus=debug
+logging.level.org.hotswap.agent.plugin.mybatisplus=debug
